Proiect_final/requirements.txt
Flask==2.2.3
google-cloud-datastore==2.15.1
google-cloud-storage==2.9.0
google-cloud-language==2.9.1
google-cloud-translate==3.11.1
